app/requirements.txt
PyQt5==5.15.9
opencv-python==4.8.0.76
Pillow==9.5.0
torch==1.8.0
torchvision==0.9.0
onnxruntime==1.15.1
paddlepaddle==2.4.2
shapely==2.0.1
pyclipper==1.3.0.post4
scikit-image==0.21.0
pymongo==4.4.1